Added the #pin feature
Unfortuantly you have to use the message id in order to pin it. I might one day if I have nothing to do, like at all find a way to find a image based on content. but I would have to get EVERY message in that channel and then find the one message. Which would not be very good for optimization.