September 18, 2021
Uh, crap I did quite a bit. Since I'm in 2 servers now one with full admin and one that is quite restrictive, have different commands for different servers as to not crash the bot, So I added that. Tag and untag in the way so every 5 messages he will say among us. Might work that out with some User ID assignment but I'm not smart enough for that so, I'll figure it out later.